Move commands into separate classes
