+++ run.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/env bash
 
 mvn clean install
+if [[ $? -ne 0 ]]; then
+    exit 1
+fi
 
 java -jar target/pdfforms.jar
