+++ src/main/java/org/pdf/forms/gui/commands/Command.java
@@ -1,0 +1,5 @@
+package org.pdf.forms.gui.commands;
+
+interface Command {
+    void execute();
+}
