+++ src/main/java/org/pdf/forms/gui/commands/ExitCommand.java
@@ -1,0 +1,10 @@
+package org.pdf.forms.gui.commands;
+
+public class ExitCommand implements Command {
+
+    @Override
+    public void execute() {
+        System.exit(0);
+    }
+
+}
