+++ src/main/java/org/pdf/forms/gui/commands/NoopCommand.java
@@ -1,0 +1,9 @@
+package org.pdf.forms.gui.commands;
+
+public class NoopCommand implements Command {
+
+    @Override
+    public void execute() {
+        // do nothing
+    }
+}
